Refactor: 회원가입 로직 컴포넌트 분리
