+++ src/components/signin/SignInForm.tsx
@@ -106,11 +106,10 @@ const SignInForm = () => {
               )}
             />
             <div className={"flex justify-between gap-2"}>
-              <Button type="submit" variant={"outline"}>
-                로그인
-              </Button>
+              <Button type="submit">로그인</Button>
               <Button
                 type="button"
+                variant={"outline"}
                 onClick={() => router.push("/signup/SignUpForm")}
               >
                 회원가입
