+++ src/components/signup/index.ts
@@ -1,4 +1,3 @@
-export { default as SignUpForm } from "./SignUpForm";
 export { default as FirstStep } from "./FirstStep";
 export { default as LastStep } from "./LastStep";
 export { default as StepButton } from "./StepButton";
